# Conflicts:
#	src/main/java/com/javamentor/qa/platform/dao/abstracts/dto/QuestionDtoDao.java
#	src/main/java/com/javamentor/qa/platform/dao/abstracts/model/ReputationDAO.java
#	src/main/java/com/javamentor/qa/platform/dao/impl/dto/QuestionDtoDaoImpl.java
#	src/main/java/com/javamentor/qa/platform/models/dto/QuestionDto.java
#	src/main/java/com/javamentor/qa/platform/service/impl/TestDataEntityService.java
#	src/main/java/com/javamentor/qa/platform/service/impl/model/ReputationServiceImpl.java
#	src/main/resources/static/js/util/add_question.js

+++ src/main/java/com/javamentor/qa/platform/dao/abstracts/dto/SearchQuestionDAO.java
@@ -12,7 +12,7 @@ public interface SearchQuestionDAO {
 
     List<QuestionDto> getQuestionsByNumberOfAnswers(Long numberOfAnswers);
 
-    List<QuestionDto> getQuestionsByNumberOfVotes(Integer numberOfVotes);
+    List<QuestionDto> getQuestionsByNumberOfVotes(Long numberOfVotes);
 
     List<QuestionDto> getQuestionsByFieldHelpfulFalse();
